Add support for libgpiod 2.x
